refactor: next 마이그레이션 - import 경로 수정
