+++ src/components/FolderAddButton.tsx
@@ -1,9 +1,9 @@
 import styled from 'styled-components';
 import addIcon from '../../../assets/addIcon.svg';
 import addWhiteIcon from '../../../assets/addWhiteIcon.svg';
-import { ModalContext } from '../FolderPage';
+import { ModalContext } from '@/pages/folder';
 import { useContext } from 'react';
-import { EDIT_TYPE, FOLDER_ADD } from '../../../constants/modalConstants';
+import { EDIT_TYPE, FOLDER_ADD } from '@/src/constants/modalConstants';
 
 //styled-components로 구현한 폴더추가 버튼입니다.
 // 기본 버튼과 모바일 버튼을 따로 구현해서 사용합니다.
