+++ src/components/Footer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../styles/footer.css';
+import '@/styles/components-styles/footer.css';
 
 function Footer() {
   return (
